# Improve predicate performance for `in` operator