@crepererum 2019-09-16
I think this is a tradeoff. If the number of variables in the predicate is rather low and especially wide-spread, so the set of matched row-groups is low, then your change would be especially effective. If the array in the predicate is very long, then your change might even be counter-productive due to all the python-driven iteration over all the predicate values. So: can we have another test case where the predicate is large?

We could probably also implement both variants (per-value filter and min-max-based filter) and select the algorithm based on the size of the predicate (e.g. if the "in"-clause contains more than 100 elements, just use min and max to filter row-groups).