@lr4d 2019-09-17
>If the array in the predicate is very long, then your change might even be counter-productive due to all the python-driven iteration over all the predicate values.

Yes, this occurred to me. I will do some more benchmarks when I have the time to check performance of larger predicates.