@lr4d 2019-09-17
I ran some benchmarks again with predicates in ranges of 2M, 20K and 3. I compared master, the `for` loop implementation and the implementation suggested by @fjetter using numpy

Predicates used:
```python
matching_long_predicate_value = tuple(range(-10**6, 10**6))
long_predicate_value = (*range(-10**6, 15), *range(20, 10**6))
mid_predicate_value = (*range(-10**4, 15), *range(20, 10**4))
short_predicate_value = (13, 15, 20)
matching_short_predicate_value = (13, 18, 20)
```

Results:
```
# Using for loop

# Long predicate which matches: (length=2M)
# 7.01 s ± 1.92 s per loop (mean ± std. dev. of 7 runs, 1 loop each)

# Long predicate (lenght=1999995):
# 6.62 s ± 672 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# Mid-sized predicate (length=19995)
# 61 ms ± 1.16 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# Short predicate:
# 3.96 ms ± 375 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# Matching short predicate:
# 75.8 ms ± 11.4 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# ------------
# Using `np.any`
# Long predicate (lenght=1999995):
#6.69 s ± 360 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# Mid-sized predicate (length=19995)
# 73.3 ms ± 5.22 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# Short predicate:
# 4.97 ms ± 793 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)


# ---------------
# Current implementation (master branch):

# Long predicate which matches: (length=2M)
# 6.24 s ± 451 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# Long predicate (lenght=1999995):
# 5.21 s ± 330 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# Mid-sized predicate (length=19995)
# 213 ms ± 15.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# Short predicate:
# 57.9 ms ± 6.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# Matching short predicate:
# 63.2 ms ± 3.57 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

```

Obviously some margin of error needs to be taken into account with these results, esp. for large predicates.

### Conclusion
If the values provided as `in` predicates are _not_ in the actual Parquet file and the current predicate logic returns True, this will be faster than the current predicate logic, the difference being relatively big for `len(values) <= 50K`, however, at `len(values)> 1M` this will actually be slower.
On the other hand, if the current predicate logic returns False, this will probably be slower in all cases (however, this difference is relatively small). 
If the values provided as  `in` predicates are in between the Parquet statistics, this will also be slower than the current predicate logic, although again the difference is relatively small.
 
Note that, in any case, when predicate values are larger than ~1M, this implementation will be slower than the current one.
Thus, I think @crepererum suggestion about using this for small predicates, and using the current implementation for large predicates makes sense.