@fjetter 2019-09-17
The results are quite striking and I cannot see any obvious errors in the benchmark setup which could account for large deviations (Sure, it's only one data type, data setup is constant, etc. but that shouldn't have a significant impact).

Therefore, I'd also be fine with the for-loop-only implementation with a dedicated comment about the performance and a link to this issue. The measurements above indicate that even for a list of 20k items the for loop performs better. A list of 20k items doesn't count as a _mid-sized_ list anymore but is so large that I would start questioning the approach of using predicate pushdown by itself (We have lists up to ~200 items in production).

Side note: sorting the list and performing some sort of binary search might improve this even further but at some point I would argue that the performance is good enough 😁